YUK-18841 Add messaging from LM->Service Manager for Nest YUKRV-12788
- Review changes - converted client messaging to 64 bit timestamps.

+++ yukon-server/LoadManagement/lmgroupnest.cpp
@@ -45,7 +45,7 @@ bool LMGroupNest::sendCycleControl( long controlDurationSeconds )
         OutboundQueue::NestCyclingControl, 
         std::make_unique<LMNestCyclingControlMessage>( 
             getPAOId(), 
-            static_cast<int>( utcNow.seconds() ),
+            utcNow.seconds(),
             controlDurationSeconds ) );
 
     if ( _LM_DEBUG & LM_DEBUG_STANDARD )
@@ -72,7 +72,7 @@ bool LMGroupNest::sendStopControl( bool stopImmediately )
         OutboundQueue::NestRestore,
         std::make_unique<LMNestRestoreMessage>(
             getPAOId(),
-            static_cast<int>( utcNow.seconds() ) ) );
+            utcNow.seconds() ) );
 
     if ( _LM_DEBUG & LM_DEBUG_STANDARD )
     {
@@ -98,7 +98,7 @@ bool LMGroupNest::sendShedControl( long controlMinutes )
         OutboundQueue::NestRestore,
         std::make_unique<LMNestRestoreMessage>(
             getPAOId(),
-            static_cast<int>( utcNow.seconds() ) ) );
+            utcNow.seconds() ) );
 
     if ( _LM_DEBUG & LM_DEBUG_STANDARD )
     {
