+++ yukon-server/Message/include/LMNestMessages.h
@@ -13,17 +13,17 @@ class IM_EX_MSG LMNestCyclingControlMessage : public StreamableMessage
 {
 public:
 
-    LMNestCyclingControlMessage( int groupId,
-                                 int startTime,
-                                 int controlDuration );
+    LMNestCyclingControlMessage( int       groupId,
+                                 long long startTime,
+                                 int       controlDuration );
 
     void streamInto( cms::StreamMessage & message ) const;
 
 private:
 
-    int _groupId;
-    int _startTime;
-    int _stopTime;
+    int       _groupId;
+    long long _startTime;
+    long long _stopTime;
 };
 
 /// 
@@ -32,15 +32,15 @@ class IM_EX_MSG LMNestRestoreMessage : public StreamableMessage
 {
 public:
 
-    LMNestRestoreMessage( int groupId,
-                          int restoreTime );
+    LMNestRestoreMessage( int       groupId,
+                          long long restoreTime );
 
     void streamInto( cms::StreamMessage & message ) const;
 
 private:
 
-    int _groupId;
-    int _restoreTime;
+    int       _groupId;
+    long long _restoreTime;
 };
 
 }
